Fix compilation on Win32/MSVC
svn path=/trunk/koffice/; revision=1140931

+++ krita/plugins/paintops/libpaintop/sensors/kis_dynamic_sensor_time.cc
@@ -36,7 +36,7 @@ double KisDynamicSensorTime::parameter(const KisPaintInformation&  pi)
     m_lastTime = pi.currentTime();
     if (m_time > m_length) {
         if (m_periodic) {
-            m_time = fmod(m_time, m_length);
+            m_time = (int)fmod((float)m_time, (float)m_length);
         } else {
             m_time = m_length;
         }
